[ext] Rename some modules to variants of "ext"
